fix: move representative photo util
