+++ src/lib/caseUtils.ts
@@ -1,0 +1,5 @@
+import type { Case } from "./caseStore";
+
+export function getRepresentativePhoto(caseData: Pick<Case, "photos">): string {
+  return [...caseData.photos].sort()[0];
+}
